requirements.txt
google-api-python-client==2.84.0
google-auth-oauthlib==1.0.0
